BigFloat compatibility for univariate optimization, fixes #271
The univariate optimization code unintentionally assumed that the choice
variable was a Float64.  Specifically, the interfaces accommodated an
AbstractFloat, but then would crash when reporting the results because of
missing typecasts.

This patch introduces approriate typecasts and test cases.

+++ src/brent.jl
@@ -162,8 +162,8 @@ function optimize{F <: Function, T <: AbstractFloat}(
                                          iteration,
                                          iteration == iterations,
                                          converged,
-                                         rel_tol,
-                                         abs_tol,
+                                         Float64(rel_tol),
+                                         Float64(abs_tol),
                                          tr,
                                          f_calls)
 end
