+++ src/golden_section.jl
@@ -103,8 +103,8 @@ function optimize{F<:Function, T <: AbstractFloat}(f::F, x_lower::T, x_upper::T,
                                          iteration,
                                          iteration == iterations,
                                          converged,
-                                         rel_tol,
-                                         abs_tol,
+                                         Float64(rel_tol),
+                                         Float64(abs_tol),
                                          tr,
                                          f_calls)
 end
